Display errors in login page
Now the app shows the reason why the login failed to the user.

Why this happened?
There's a 'Consumer' at the top most widget (MaterialApp) which gets rebuild everytime during logout and login. We don't need to notify listeners and rebuild the MaterialApp widget when logging out (cleaning up old data) during login. By doing so, context is in place and dialog appears. If the whole MaterialApp rebuilds, the old context is gone and a new context is given to all the widgets, which is where the issue arises and the dialog uses old context (which doens't exist in the first place).

Signed-off-by: AxelBlaz3 <karthikgaddam4@gmail.com>